refactor(chain): split finalized RPC transport boundaries
